+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added method overview in documentation
+- `StableVec::contains()`
 
 
 ## [0.1.2] - 2017-09-15